Ifm look

+++ app/controllers/bookings_controller.rb
@@ -1,23 +1,37 @@
 class BookingsController < ApplicationController
   def index
+    redirect_to admin_bookings_path if current_user.admin?
     @bookings = current_user.bookings.in_the_future
   end
 
   def create
+    booking = Booking.new(booking_params)
+    booking.machine_id = params[:machine_id]
+    booking.user = current_user
+
     unless current_user.admin?
-      if current_user.bookings.in_the_future.count > 2
+      if current_user.formations.where(machine: booking.machine).count.zero?
+        flash[:alert] = "looks like you haven't passed the formation for this machine..."
+        return redirect_to root_path
+      end
+      if current_user.bookings.in_the_future.count > 10
         flash[:alert] = 'already too much bookings'
+        return redirect_to root_path
+      end
+      if booking.start_time + 1.hours != booking.end_time || booking.start_time > (DateTime.now + 74.hours)
+        flash[:alert] = 'Nice try.'
+        return redirect_to root_path
+      end
+      unless on_that_day(booking)
+        flash[:alert] = "You already book the #{booking.machine.name} twice this day"
+        return redirect_to root_path
+      end
+      unless current_user.bookings.where(start_time: booking.start_time).empty?
+        flash[:alert] = "Looks like you already booked #{current_user.bookings.where(start_time: booking.start_time).last.machine.name} that day at #{booking.start_time.strftime("%H:00")}.. You can't book two machines at the same time."
         return redirect_to root_path
       end
     end
 
-    booking = Booking.new(booking_params)
-    booking.machine_id = params[:machine_id]
-    booking.user = current_user
-    unless current_user.bookings.where(start_time: booking.start_time).empty?
-      flash[:alert] = "Looks like you already booked #{current_user.bookings.where(start_time: booking.start_time).last.machine.name} that day at #{booking.start_time.strftime("%H:00")}.. You can't book two machines at the same time."
-      return redirect_to root_path
-    end
     if booking.save!
       flash[:notice] = 'booking created'
     else
@@ -28,8 +42,12 @@ class BookingsController < ApplicationController
 
   def destroy
     booking = Booking.find(params[:id])
+    unless current_user.admin?
+      return redirect_to root_path if booking.user != current_user
+    end
+
     booking.destroy
-    redirect_to bookings_path
+    redirect_to machine_path(booking.machine)
     flash[:notice] = 'booking was succesfully cancelled'
   end
